+++ src/types/People.ts
@@ -1,3 +1,8 @@
+import { Link } from "react-router-dom";
+
 export default interface PeopleData {
   name: string;
+  username: string;
+  professionalHeadline: string;
+  picture: string;
 }