+++ CHANGELOG.md
@@ -1,11 +1,40 @@
 # Changelog
 
+## 2021-09-19 02:00 GTM-5
+
+Changes:
+
+- Add infinite scroll: hooks control, conditional loader and control of data to show
+- Customize empty, start and end message when fetching data
+- Size param static in API = 20
+
+Bugs:
+
+- Wrong number of results in [https://torre.co/search/people-name?q=name%3Atorrenegra](https://torre.co/search/people-name?q=name%3Atorrenegra)
+
+
+<p align="center">
+  <a href="#">
+    <img src="https://imgur.com/5JuSyJx.png" width="500">
+  </a>
+</p>
+
+> 370 results on torre.co
+
+<p align="center">
+  <a href="#">
+    <img src="https://imgur.com/YysGPUC.png" width="600">
+  </a>
+</p>
+
+> 37 results on API
+
 ## 2021-09-18 19:00 GTM-5
 
 Changes:
 
 - Add search bar with events
-- Support for queries with after API pagination 
+- Support for queries with after API pagination
 
 ## 2021-09-18 18:00 GTM-5